+++ src/collector/CollectorWrapper.test.ts
@@ -2,18 +2,17 @@ import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { mockWindowLocation, mockWindowScreen } from '../test-utils/mocks'
 import ClickEventListener from '../event/listener/ClickEventListener'
 import ChangeEventListener from '../event/listener/ChangeEventListener'
-import { ConsoleEventHandler } from '../event/handler/ConsoleEventHandler'
 import CollectorWrapper from './CollectorWrapper'
-import { createSessionEvent } from '../event/createSessionEvent'
+import { createSessionStartedEvent } from '../event/createSessionStartedEvent'
 import { CollectorOptions } from '../types'
+import UnloadEventListener from '../event/listener/UnloadEventListener'
+import EventHandler from '../event/handler/EventHandler'
 
 describe('CollectorWrapper', () => {
   beforeEach(() => {
     mockWindowScreen()
     mockWindowLocation()
-    vi.spyOn(ConsoleEventHandler.prototype, 'run').mockImplementation(() => {
-      return {}
-    })
+    vi.spyOn(EventHandler.prototype, 'run').mockImplementation(() => {})
   })
 
   describe('constructor', () => {
@@ -27,12 +26,8 @@ describe('CollectorWrapper', () => {
 
     describe('when debug option is set to true', () => {
       beforeEach(() => {
+        // @ts-expect-error
         options = { debug: true }
-      })
-
-      it('instantiates a ConsoleEventHandler by default', () => {
-        const sut = createCollectorWrapper()
-        expect(sut.eventHandler).toBeInstanceOf(ConsoleEventHandler)
       })
 
       it('a "sessionStarted" event is sent when initialized', () => {
@@ -40,56 +35,33 @@ describe('CollectorWrapper', () => {
         vi.useFakeTimers()
         vi.setSystemTime(Date.parse('2022-05-12'))
 
-        const expectedEvent = createSessionEvent()
+        const expectedEvent = createSessionStartedEvent()
 
-        const mock = vi.spyOn(ConsoleEventHandler.prototype, 'run').mockImplementation(() => {
-          return {}
-        })
+        const mock = vi.spyOn(EventHandler.prototype, 'run').mockImplementation(() => {})
 
         createCollectorWrapper()
         expect(mock).toHaveBeenCalledWith(expectedEvent)
       })
 
       it('initializes ClickEventListener', () => {
-        vi.spyOn(ClickEventListener.prototype, 'init').mockImplementation(() => {
-          return {}
-        })
+        vi.spyOn(ClickEventListener.prototype, 'init').mockImplementation(() => {})
         createCollectorWrapper()
 
         expect(ClickEventListener.prototype.init).toHaveBeenCalledOnce()
       })
 
       it('initializes ChangeEventListener', () => {
-        vi.spyOn(ChangeEventListener.prototype, 'init').mockImplementation(() => {
-          return {}
-        })
+        vi.spyOn(ChangeEventListener.prototype, 'init').mockImplementation(() => {})
         createCollectorWrapper()
 
         expect(ChangeEventListener.prototype.init).toHaveBeenCalledOnce()
       })
-    })
 
-    describe('when debug option is set to false (default version)', () => {
-      beforeEach(() => {
-        options = { debug: false }
-      })
+      it('initializes UnloadEventListener', () => {
+        vi.spyOn(UnloadEventListener.prototype, 'init').mockImplementation(() => {})
+        createCollectorWrapper()
 
-      it('throws an error if AuthKey is not set', () => {
-        expect(() => createCollectorWrapper()).toThrowError('No AuthKey was specified')
-      })
-
-      describe('when AuthKey is specified', () => {
-        beforeEach(() => {
-          options = {
-            debug: false,
-            authKey: '123-456-789',
-          }
-        })
-
-        it('throws an error', () => {
-          // TODO: use a real eventCollector would be way nicer ;)
-          expect(() => createCollectorWrapper()).toThrowError('Not implemented yet')
-        })
+        expect(UnloadEventListener.prototype.init).toHaveBeenCalledOnce()
       })
     })
   })