+++ src/event/createGravityEvent.test.ts
@@ -1,10 +1,10 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { mockClick, mockWindowLocation, mockWindowScreen } from '../test-utils/mocks'
-import { JSDOM } from 'jsdom'
 import { createGravityEvent } from './createGravityEvent'
 import viewport from '../utils/viewport'
 import location from '../utils/location'
 import { EventType } from '../types'
+import createElementInJSDOM from '../test-utils/createElementInJSDOM'
 
 describe('event', () => {
   beforeEach(() => {
@@ -19,25 +19,25 @@ describe('event', () => {
     })
 
     it('returns the specified "type"', async () => {
-      const dom = new JSDOM('<button>Click Me</button>')
-      const element = dom.window.document.querySelector('button')
-      const event = await createGravityEvent(mockClick(element as HTMLElement) as unknown as Event, EventType.Click)
+      const { element } = createElementInJSDOM('<button>Click Me</button>', 'button')
+
+      const event = await createGravityEvent(mockClick(element), EventType.Click)
 
       expect(event.type).toEqual(EventType.Click)
     })
 
     it('returns location data', async () => {
-      const dom = new JSDOM('<div>Click Me</div>')
-      const element = dom.window.document.querySelector('div')
-      const event = await createGravityEvent(mockClick(element as HTMLElement) as unknown as Event, EventType.Click)
+      const { element } = createElementInJSDOM('<div>Click Me</div>', 'div')
+
+      const event = await createGravityEvent(mockClick(element), EventType.Click)
 
       expect(event.location).toEqual(location())
     })
 
     it('returns viewport data', async () => {
-      const dom = new JSDOM('<div>Click Me</div>')
-      const element = dom.window.document.querySelector('div')
-      const event = await createGravityEvent(mockClick(element as HTMLElement) as unknown as Event, EventType.Click)
+      const { element } = createElementInJSDOM('<div>Click Me</div>', 'div')
+
+      const event = await createGravityEvent(mockClick(element), EventType.Click)
 
       expect(event.viewportData).toEqual(viewport())
     })
@@ -47,34 +47,35 @@ describe('event', () => {
       vi.useFakeTimers()
       vi.setSystemTime(Date.parse('2022-05-12'))
 
-      const dom = new JSDOM('<div>Click Me</div>')
-      const element = dom.window.document.querySelector('div')
-      const event = await createGravityEvent(mockClick(element as HTMLElement) as unknown as Event, EventType.Click)
+      const { element } = createElementInJSDOM('<div>Click Me</div>', 'div')
+
+      const event = await createGravityEvent(mockClick(element), EventType.Click)
 
       expect(event.recordedAt).toEqual(now)
     })
 
     describe('target data embeds...', () => {
       it('tag name', async () => {
-        const dom = new JSDOM('<div>Click Me</div>')
-        const element = dom.window.document.querySelector('div')
-        const event = await createGravityEvent(mockClick(element as HTMLElement) as unknown as Event, EventType.Click)
+        const { element } = createElementInJSDOM('<div>Click Me</div>', 'div')
+
+        const event = await createGravityEvent(mockClick(element), EventType.Click)
 
         expect(event.target?.element).toEqual('div')
       })
 
       it('text content', async () => {
-        const dom = new JSDOM('<li>I am a list item</li>')
-        const element = dom.window.document.querySelector('li')
-        const event = await createGravityEvent(mockClick(element as HTMLElement) as unknown as Event, EventType.Click)
+        const { element } = createElementInJSDOM('<li>I am a list item</li>', 'li')
+        if (element == null) throw new Error('Element is null')
+
+        const event = await createGravityEvent(mockClick(element), EventType.Click)
 
         expect(event.target?.textContent).toBeUndefined()
       })
 
       it('html attributes', async () => {
-        const dom = new JSDOM('<input type="text" data-testid="userName" class="size-lg"/>')
-        const element = dom.window.document.querySelector('input')
-        const event = await createGravityEvent(mockClick(element as HTMLElement) as unknown as Event, EventType.Click)
+        const { element } = createElementInJSDOM('<input type="text" data-testid="userName" class="size-lg"/>', 'input')
+
+        const event = await createGravityEvent(mockClick(element), EventType.Click)
 
         expect((event.target?.attributes as Record<string, string>).type).toEqual('text')
         expect((event.target?.attributes as Record<string, string>)['data-testid']).toEqual('userName')
@@ -82,9 +83,9 @@ describe('event', () => {
       })
 
       it('pointer coordinates data when the event is a click', async () => {
-        const dom = new JSDOM('<input type="text" data-testid="userName" class="size-lg"/>')
-        const element = dom.window.document.querySelector('input')
-        const clickEvent = mockClick(element as HTMLElement) as unknown as PointerEvent
+        const { element } = createElementInJSDOM('<input type="text" data-testid="userName" class="size-lg"/>', 'input')
+
+        const clickEvent = mockClick(element)
         const event = await createGravityEvent(clickEvent, EventType.Click)
 
         expect(event.eventData?.clickOffsetX).toEqual(clickEvent.clientX)