+++ src/event/handler/IEventHandler.ts
@@ -1,5 +1,0 @@
-import { TEvent } from '../../types'
-
-export default interface IEventHandler {
-  run: (event: TEvent) => void
-}