+++ src/event/listener/ChangeEventListener.test.ts
@@ -1,63 +1,60 @@
-import sinon from 'sinon'
-import { afterEach, beforeEach, describe, expect, it } from 'vitest'
-import { JSDOM } from 'jsdom'
-import IEventHandler from '../handler/IEventHandler'
+import { beforeEach, describe, expect, it, vitest } from 'vitest'
+import EventHandler from '../handler/EventHandler'
 import { fireEvent, getByRole, waitFor } from '@testing-library/dom'
 import ChangeEventListener from './ChangeEventListener'
+import { nop } from '../../utils/nop'
+import createElementInJSDOM from '../../test-utils/createElementInJSDOM'
 
 describe('ChangeEventListener', () => {
   describe('listener', () => {
-    let eventHandler: IEventHandler
-    let runSpy: sinon.SinonSpy
+    const eventHandler = new EventHandler('aaa-111', 0, nop)
+    const runSpy = vitest.spyOn(eventHandler, 'run')
 
     beforeEach(() => {
-      eventHandler = {
-        run: () => {},
-      }
-      runSpy = sinon.spy(eventHandler, 'run')
-    })
-
-    afterEach(() => {
-      runSpy.restore()
+      vitest.restoreAllMocks()
     })
 
     it('calls listener when change event been fired', async () => {
-      const dom = new JSDOM(`
-                <div>
-                    <input type="checkbox" id="checkbox1" name="checkbox1">
-                </div>`)
+      const { element, domWindow } = createElementInJSDOM(
+        `
+        <div>
+          <input type='checkbox' id='checkbox1' name='checkbox1'>
+        </div>`,
+        'div',
+      )
 
-      new ChangeEventListener(eventHandler, dom.window as unknown as Window).init()
+      new ChangeEventListener(eventHandler, domWindow).init()
 
-      const container = dom.window.document.querySelector('div')
-      const checkBox = await waitFor(() => getByRole(container as HTMLElement, 'checkbox'))
+      const checkBox = await waitFor(() => getByRole(element, 'checkbox'))
 
       fireEvent.change(checkBox)
 
       await waitFor(() => {
-        expect(runSpy.calledOnce).toBeTruthy()
+        expect(runSpy).toHaveBeenCalledOnce()
       })
     })
 
     it('does not call listener if target is not a checkbox', async () => {
-      const dom = new JSDOM(`
-                <div>
-                    <input type="text" class="size-lg"/>
-                </div>`)
+      const { element, domWindow } = createElementInJSDOM(
+        `
+        <div>
+            <input type='text' class='size-lg'/>
+        </div>`,
+        'div',
+      )
 
-      const listener = new ChangeEventListener(eventHandler, dom.window as unknown as Window)
-      const listenerSpy = sinon.spy(listener, 'listener')
+      const listener = new ChangeEventListener(eventHandler, domWindow)
+      const listenerSpy = vitest.spyOn(listener, 'listener')
 
       listener.init()
 
-      const container = dom.window.document.querySelector('div')
-      const input = await waitFor(() => getByRole(container as HTMLElement, 'textbox'))
+      const input = await waitFor(() => getByRole(element, 'textbox'))
 
       fireEvent.change(input)
 
       await waitFor(() => {
-        expect(listenerSpy.calledOnce).toBeTruthy()
-        expect(runSpy.notCalled).toBeTruthy()
+        expect(listenerSpy).toHaveBeenCalledOnce()
+        expect(runSpy).toHaveBeenCalledTimes(0)
       })
     })
   })