+++ vite.config.ts
@@ -5,5 +5,6 @@ export default defineConfig({
     globals: true,
     environment: 'jsdom',
     restoreMocks: true,
+    setupFiles: ['./src/setupTests.ts'],
   },
 })